requirements.txt
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.4.2
pyyaml==5.3.1
pyopenssl=19.1.0
google-auth=1.12.0
